initial commit
AddScore(int addingScore) method added to MainManager script and we keep score in MainManager. We call this method when triggered with platforms and ring.When we triggered with ring adding +10 point, when triggered with platform adding -1 point. For prevent case of triggiring 2 platform at same time and because of this add-2 score , we created a boolean "losePoint" and if statament . Also scoreText is working.